Assigned a class to each landing page
In the frame of this PR I also tried to provide a better name to managers.
The rationale of this PR is to assign to each section of each area of surveypro a specific class.
Once each section will use its specific class it will be simple to:
- better debug the code;
- make code improvements;
- create unit tests.

As far as I know, this PR along with #932 and #931, accomplishes this goal.

What was in my mind (and I hope I reached this target) was to assign a specific class to each landing page of sPro.
In other words...
sPro is divided into 6 areas and each area is divided into sections.
The structure of sPro, is:

[Spro area] surveypro (managed in view.php)
- [Area section] cover
- [Area section] submissionslist
- [Area section] submissionform
- [Area section] searchsubmissions
[Spro area] layout (managed in layout.php)
- [Area section] itemslist
- [Area section] itemsetup
- [Area section] branchingvalidation
- [Area section] preview
[Spro area] reports (managed in reports.php)
- [report] attachement overview
- [report] frequency distribution
- [report] late users
- [report] responses per user
- [report] users per count of responses
[Spro area] tools (managed in tools.php)
- [Area section] export
- [Area section] import
[Spro area] user templates (managed in utemplates.php)
- [Area section] manage
- [Area section] save
- [Area section] import
- [Area section] apply
[Spro area] master templates (managed in mtemplates.php)
- [Area section] save
- [Area section] apply

Well, if I am not wrong, now each section uses a specific class. (reports are an exception)

Names of classes always start with the name of the area and ends with the name of the section.
They are:

[Spro area] surveypro (managed in view.php)
- [Area section] cover -----------------------------------> view_cover
- [Area section] submissionslist -------------------------> view_submissionlist
- [Area section] submissionform --------------------------> view_submissionform
- [Area section] searchsubmissions -----------------------> view_submissionsearch
[Spro area] layout (managed in layout.php)
- [Area section] itemslist -------------------------------> layout_itemlist
- [Area section] itemsetup -------------------------------> layout_itemsetup
- [Area section] branchingvalidation ---------------------> layout_branchingvalidation
- [Area section] preview ---------------------------------> layout_preview
[Spro area] reports (managed in reports.php)
- [report] attachement overview --------------------------> report from its own specific plugin classes
- [report] frequency distribution ------------------------> report from its own specific plugin classes
- [report] late users ------------------------------------> report from its own specific plugin classes
- [report] responses per user ----------------------------> report from its own specific plugin classes
- [report] users per count of responses ------------------> report from its own specific plugin classes
[Spro area] tools (managed in tools.php)
- [Area section] export -----------------------------------> tools_export
- [Area section] import -----------------------------------> tools_import
[Spro area] user templates (managed in utemplates.php)
- [Area section] manage -----------------------------------> utemplate_manage
- [Area section] save -------------------------------------> utemplate_save
- [Area section] import -----------------------------------> utemplate_import
- [Area section] apply ------------------------------------> utemplate_apply
[Spro area] master templates (managed in mtemplates.php)
- [Area section] save -------------------------------------> mtemplate_save
- [Area section] apply ------------------------------------> mtemplate_apply

The corresponding manager (still, if I am not wrong) is named with the name of the section and finishes with 'man'.
So, for instance: corverman, itemlistman, exportman, manageman and so forth.
